Add audio

+++ Goomba.h
@@ -25,7 +25,7 @@ public:
 	void OnCollisionWith(CollisionEvent* event) override;
 	Rect GetBoundingBox() override
 	{
-		return Rect::FromXYWH(position.x, position.y, 16, 16);
+		return Rect::FromXYWH(static_cast<int>(position.x), static_cast<int>(position.y), 16, 16);
 	}
 };
